--- docs/components/Footer.vue
@@ -1,14 +1,0 @@
-<template>
-    <footer class="w-full py-4 bg-slate-200 border-t shadow-sm">
-        <div class="w-8/12 mx-auto flex justify-between">
-            <div>
-                &copy; 2021 Enchant-scrollbar. All rights reserved.
-            </div>
-
-            <div>
-                <a href="#" class="text-slate-600">Privacy Policy</a>
-                <a href="#" class="text-slate-600 ml-4">Terms of Service</a>
-            </div>
-        </div>
-    </footer>
-</template>
